+++ app/src/main/kotlin/org/eurofurence/connavigator/broadcast/LogoutReceiver.kt
@@ -3,7 +3,7 @@ package org.eurofurence.connavigator.broadcast
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import org.eurofurence.connavigator.gcm.InstanceIdService
+import org.eurofurence.connavigator.services.InstanceIdService
 import org.eurofurence.connavigator.pref.AuthPreferences
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
