+++ app/src/main/kotlin/org/eurofurence/connavigator/broadcast/ResetReceiver.kt
@@ -6,10 +6,9 @@ import android.content.Intent
 import nl.komponents.kovenant.task
 import org.eurofurence.connavigator.R
 import org.eurofurence.connavigator.database.RootDb
-import org.eurofurence.connavigator.net.imageService
+import org.eurofurence.connavigator.services.ImageService
 import org.eurofurence.connavigator.pref.AnalyticsPreferences
 import org.eurofurence.connavigator.pref.AppPreferences
-import org.eurofurence.connavigator.pref.AuthPreferences
 import org.eurofurence.connavigator.pref.DebugPreferences
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
@@ -30,7 +29,7 @@ class ResetReceiver : BroadcastReceiver(), AnkoLogger {
         RootDb(context).clear()
 
         info { "Purging images" }
-        imageService.clear()
+        ImageService.clear()
 
         info { "Bombarding user settings" }
         AppPreferences.clear()
