+++ app/src/main/kotlin/org/eurofurence/connavigator/services/InstanceIdService.kt
@@ -1,4 +1,4 @@
-package org.eurofurence.connavigator.gcm
+package org.eurofurence.connavigator.services
 
 import com.google.firebase.iid.FirebaseInstanceId
 import com.google.firebase.iid.FirebaseInstanceIdService
@@ -6,7 +6,6 @@ import io.swagger.client.model.PostFcmDeviceRegistrationRequest
 import nl.komponents.kovenant.task
 import org.eurofurence.connavigator.BuildConfig
 import org.eurofurence.connavigator.pref.AuthPreferences
-import org.eurofurence.connavigator.webapi.apiService
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
 import org.jetbrains.anko.warn
