+++ app/src/main/kotlin/org/eurofurence/connavigator/ui/fragments/EventRecyclerFragment.kt
@@ -5,7 +5,6 @@ import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.fragment.app.Fragment
@@ -23,7 +22,7 @@ import org.eurofurence.connavigator.R
 import org.eurofurence.connavigator.database.HasDb
 import org.eurofurence.connavigator.database.glyphsFor
 import org.eurofurence.connavigator.database.lazyLocateDb
-import org.eurofurence.connavigator.net.imageService
+import org.eurofurence.connavigator.services.ImageService
 import org.eurofurence.connavigator.ui.dialogs.eventDialog
 import org.eurofurence.connavigator.ui.filters.FilterChain
 import org.eurofurence.connavigator.ui.filters.start
@@ -155,7 +154,7 @@ class EventRecyclerFragment : Fragment(), HasDb, AnkoLogger {
                 backgroundColorResource = R.color.cardview_light_background
                 visibility = View.GONE
                 from { bannerImageId } into {
-                    imageService.load(db.images[it], this)
+                    ImageService.load(db.images[it], this)
                 }
             }.lparams(matchParent, wrapContent) {
                 topMargin = dip(-10)
